configure Git; minor refactoring
